Split filer
